+++ angular-workspace/projects/ni/nimble-angular/src/directives/drawer/nimble-drawer.directive.ts
@@ -1,40 +1,31 @@
-import { Directive, ElementRef, EventEmitter, HostListener, Input, Output, Renderer2 } from '@angular/core';
-import type { Drawer } from '@ni/nimble-components/dist/esm/drawer';
-import { DrawerLocation, DrawerState } from '@ni/nimble-components/dist/esm/drawer/types';
+import { Directive, ElementRef, Input, Renderer2 } from '@angular/core';
+import type { Drawer, UserDismissed } from '@ni/nimble-components/dist/esm/drawer';
+import { DrawerLocation } from '@ni/nimble-components/dist/esm/drawer/types';
 import { BooleanValueOrAttribute, toBooleanProperty } from '../utilities/template-value-helpers';
 
 export type { Drawer };
-export { DrawerLocation, DrawerState };
+export { DrawerLocation };
 
 /**
- * Directive to provide Angular integration for the drawer.
+ * Directive to provide Angular integration for the drawer element.
+ * @description
+ * In order to provide a custom value for the `CloseReason` for the drawer you need to obtain
+ * a `@ViewChild` reference with the custom type specified. For example, to specify the type as `string`:
+ * ```ts
+ * @ViewChild('drawer', { read: NimbleDrawerDirective }) public drawer: NimbleDrawerDirective<string>;
+ * ```
  */
 @Directive({
     selector: 'nimble-drawer'
 })
-export class NimbleDrawerDirective {
+// eslint-disable-next-line @typescript-eslint/no-invalid-void-type
+export class NimbleDrawerDirective<CloseReason = void> {
     public get location(): DrawerLocation {
         return this.elementRef.nativeElement.location;
     }
 
     @Input() public set location(value: DrawerLocation) {
         this.renderer.setProperty(this.elementRef.nativeElement, 'location', value);
-    }
-
-    public get state(): DrawerState {
-        return this.elementRef.nativeElement.state;
-    }
-
-    @Input() public set state(value: DrawerState) {
-        this.renderer.setProperty(this.elementRef.nativeElement, 'state', value);
-    }
-
-    public get modal(): boolean {
-        return this.elementRef.nativeElement.modal;
-    }
-
-    @Input() public set modal(value: BooleanValueOrAttribute) {
-        this.renderer.setProperty(this.elementRef.nativeElement, 'modal', toBooleanProperty(value));
     }
 
     public get preventDismiss(): boolean {
@@ -47,21 +38,27 @@ export class NimbleDrawerDirective {
         this.renderer.setProperty(this.elementRef.nativeElement, 'preventDismiss', toBooleanProperty(value));
     }
 
-    @Output() public stateChange = new EventEmitter<DrawerState>();
-    public constructor(private readonly renderer: Renderer2, private readonly elementRef: ElementRef<Drawer>) {}
-
-    public show(): void {
-        this.state = DrawerState.opening;
+    public get ariaLabel(): string | null {
+        return this.elementRef.nativeElement.ariaLabel;
     }
 
-    public hide(): void {
-        this.state = DrawerState.closing;
+    // ariaLabel property intentionally maps to the aria-label attribute
+    // eslint-disable-next-line @angular-eslint/no-input-rename
+    @Input('aria-label') public set ariaLabel(value: string | null) {
+        this.renderer.setProperty(this.elementRef.nativeElement, 'ariaLabel', value);
     }
 
-    @HostListener('state-change', ['$event'])
-    public onStateChanged($event: Event): void {
-        if ($event.target === this.elementRef.nativeElement) {
-            this.stateChange.emit(this.state);
-        }
+    public get open(): boolean {
+        return this.elementRef.nativeElement.open;
+    }
+
+    public constructor(private readonly renderer: Renderer2, private readonly elementRef: ElementRef<Drawer<CloseReason>>) {}
+
+    public async show(): Promise<CloseReason | UserDismissed> {
+        return this.elementRef.nativeElement.show();
+    }
+
+    public close(reason: CloseReason): void {
+        this.elementRef.nativeElement.close(reason);
     }
 }