+++ packages/nimble-components/src/dialog/styles.ts
@@ -14,9 +14,9 @@ import {
     subtitleFontColor
 } from '../theme-provider/design-tokens';
 import {
-    modalBackdropColorThemeColor,
-    modalBackdropColorThemeDark,
-    modalBackdropColorThemeLight
+    modalBackdropColorThemeColorStatic,
+    modalBackdropColorThemeDarkStatic,
+    modalBackdropColorThemeLightStatic
 } from '../theme-provider/design-tokens-static';
 import { Theme } from '../theme-provider/types';
 import { themeBehavior } from '../utilities/style/theme';
@@ -132,7 +132,7 @@ export const styles = css`
         Theme.light,
         css`
             dialog::backdrop {
-                background: ${modalBackdropColorThemeLight};
+                background: ${modalBackdropColorThemeLightStatic};
             }
         `
     ),
@@ -140,7 +140,7 @@ export const styles = css`
         Theme.dark,
         css`
             dialog::backdrop {
-                background: ${modalBackdropColorThemeDark};
+                background: ${modalBackdropColorThemeDarkStatic};
             }
         `
     ),
@@ -148,7 +148,7 @@ export const styles = css`
         Theme.color,
         css`
             dialog::backdrop {
-                background: ${modalBackdropColorThemeColor};
+                background: ${modalBackdropColorThemeColorStatic};
             }
         `
     )